% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3inferr-package}
\alias{mlr3inferr}
\alias{mlr3inferr-package}
\title{mlr3inferr: Inference on the Generalization Error}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Confidence interval and resampling methods for inference on the generalization error.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3inferr.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3inferr}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3inferr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Fischer \email{sebf.fischer@gmail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})

Authors:
\itemize{
  \item Sebastian Fischer \email{sebf.fischer@gmail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})
  \item Hannah Schulz-Kümpel \email{hannah.kuempel@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0003-3972-8392}{ORCID})
}

}
