% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3inference-package}
\alias{mlr3inference}
\alias{mlr3inference-package}
\title{mlr3inference: Inference on the Generalisation Error}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3inference.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3inference}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3inference/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Fischer \email{sebf.fischer@gmail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})

Authors:
\itemize{
  \item Hannah Schulz-Kümpel \email{hannah.kuempel@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0003-3972-8392}{ORCID})
}

}
